+++ lib/commanded/middleware/extract_aggregate_identity.ex
@@ -27,19 +27,24 @@ defmodule Commanded.Middleware.ExtractAggregateIdentity do
 
   # extract identity using a user defined function
   defp extract_aggregate_uuid(%Pipeline{command: command, identity: identity})
-    when is_function(identity)
-  do
+       when is_function(identity) do
     identity.(command)
   end
 
   # extract identity using a field in the command
   defp extract_aggregate_uuid(%Pipeline{command: command, identity: identity})
-    when is_atom(identity)
-  do
+       when is_atom(identity) do
     Map.get(command, identity)
   end
 
   defp prefix(aggregate_uuid, %Pipeline{identity_prefix: nil}), do: aggregate_uuid
+
+  # extract identity using a user defined function
+  defp prefix(aggregate_uuid, %Pipeline{identity_prefix: identity_prefix})
+       when is_function(identity_prefix) do
+    identity_prefix.() <> aggregate_uuid
+  end
+
   defp prefix(aggregate_uuid, %Pipeline{identity_prefix: identity_prefix}),
     do: identity_prefix <> aggregate_uuid
 end